Add files via upload
Job portal service was developed for creating an interactive job vacancy form for
candidates. This web application manage updates both from the job seekers as well
as the companies. It’s unique development methodology helps in acquiring the client
and candidate information and separating them according to the job requirements
and vacancies.

The online access to it provides details of the job. An employer being registered in
the web site has the facility to use the services. Being an authorized user he can
publish vacancy details and can search no of Employees on portal and also he can
search candidates on basis of the key skill which employee provides on registration.